requirements.txt
blinker==1.8.2
bson==0.5.10
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.0.4
zipp==3.20.2
